requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
httpx==0.24.0
sqlalchemy==2.0.22
alembic==1.11.1
pydantic==1.10.11
python-dotenv==1.0.0
twilio==8.3.0
openai==1.0.0
apscheduler==3.10.1
jinja2==3.1.2
aiofiles==23.1.0
